django-rest-first/requirements.txt
Django==3.2.5
djangorestframework==3.12.4
django-filter==2.4.0
mock==4.0.3
Pillow==8.3.1
coverage